homework_06/pyproject.toml
[tool.poetry]
name = "homework-06"
version = "0.1.0"
description = ""
authors = ["Anton Lantukh <lantukh@bk.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.0"
flask-migrate = "^4.0.5"
flask-sqlalchemy = "^3.1.1"
psycopg = "^3.1.16"
gunicorn = "^21.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
